backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
tensorflow==2.13.0
scikit-learn==1.3.2
pandas==2.1.4
# Additional dependencies to resolve conflicts
typing-extensions==4.8.0
protobuf==4.25.1

# LangChain and Autonomous Agent dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-google-genai==0.0.5
google-generativeai==0.3.2
python-dotenv==1.0.0
apscheduler==3.10.4
web3==6.11.3

# SMS/Twilio Integration
twilio==8.10.0
phonenumbers==8.13.26.4
matplotlib==3.8.2
seaborn==0.13.0
requests==2.31.0
aiofiles==23.2.1